Add support for Accumulo
This changeset is kinda hacky, but seems like the quickest way to
support Accumulo in Vagrant and AWS (with AMIs).

Leaders have a cron job wired up that runs every minute. That job waits
for the HDFS NameNode to report greater than zero DataNodes. Once that
occurs, it initializes Accumulo within HDFS and starts the required
services.

Followers have a different cron job that also runs every minute. That
job waits for Accumulo to be initialized by looking for the `/accumulo`
directory within HDFS. Once that directory exists, the Accumulo Tablet
Server service is initialized.