Fix race condition in Zookeeper DNS propagation
The DNS record for `zookeeper.service.geotrellis-spark.internal` is
referenced almost immediately (before Route 53 propagation) by the
`MesosLeader`, which causes an empty DNS record to be cached for 900
seconds.

This change creates a record in the local `hosts` file that points to
itself (via private IP), which negates the need for a remote DNS query.