Thread AWS region through to Packer template
Other steps in the deployment process depend on an `AWS_DEFAULT_REGION`
environment variable to determine the target AWS region. The AWS region
used by Packer to produce AMIs was hardcoded to use `us-east-1`. This
change threads the `AWS_DEFAULT_REGION` environment variable through to
Packer so that it gets used.

Attempts to resolve #22.